update to 0.5 alpha
When I started I couldn't think of a good stopping point to commit, so here's a
big blob of changes that turn it from an empty rails project into something
functional. Check the README for what's functional, and again, please don't fork
or submit pull requests unless it's a simple to implement change, this is
still an alpha and just there for people to stare at (and probably criticize).
Downvoting is the next commit.