add pages
The browse and latest sections now have page controls at the bottom that let you
move back and forth and jump to different pages. There was already a small bit
of page logic in the Quotes controller, some more things were added and
refactored.

Also in this release is a tweak on the Votes model that should speed up lookups
by checking the time of the vote at the database level instead of testing it in
the Rails model.